[Writer] Add a strict mode for writing
