+++ lib/nanaimo/writer.rb
@@ -3,7 +3,23 @@ module Nanaimo
   # string representation.
   #
   class Writer
+    autoload :PBXProjWriter, 'nanaimo/writer/pbxproj'
     autoload :XMLWriter, 'nanaimo/writer/xml'
+
+    # Raised when attempting to write a plist containing an object of an
+    # unsupported type.
+    #
+    class UnsupportedPlistTypeError < Error
+      def initialize(plist_format, object)
+        @plist_format = plist_format
+        @object = object
+      end
+
+      def to_s
+        "Unable to write #{@object.inspect}. " \
+          "`#{@object.class}` is an invalid object type to serialize in a #{@plist_format} plist."
+      end
+    end
 
     # The magic comment that denotes a UTF8-encoded plist.
     #
@@ -14,12 +30,13 @@ module Nanaimo
     #                         spaces and newlines to make output more legible
     # @param output [#<<] The output stream to write the plist to
     #
-    def initialize(plist, pretty = true, output = ::String.new)
+    def initialize(plist, pretty: true, output: ::String.new, strict: true)
       @plist = plist
       @pretty = pretty
       @output = output
       @indent = 0
       @newlines = true
+      @strict = strict
     end
 
     # Writes the plist to the given output.
@@ -30,10 +47,14 @@ module Nanaimo
       write_newline
     end
 
-    attr_reader :indent, :pretty, :output, :newlines
-    private :indent, :pretty, :output, :newlines
+    attr_reader :indent, :pretty, :output, :newlines, :strict
+    private :indent, :pretty, :output, :newlines, :strict
 
     private
+
+    def plist_format
+      :ascii
+    end
 
     def write_utf8
       output << UTF8
@@ -53,17 +74,26 @@ module Nanaimo
         write_array(object)
       when Dictionary, ::Hash
         write_dictionary(object)
-      when %r{[^\w\./]}, QuotedString, ''
+      when QUOTED_STRING_REGEXP, QuotedString, ''
         write_quoted_string(object)
       when String, ::String, Symbol
         write_string(object)
       when Data
         write_data(object)
       else
-        raise "Cannot write #{object.inspect} to an ascii plist"
+        raise UnsupportedPlistTypeError.new(plist_format, object) if strict
+        write_string_quoted_if_necessary(object)
       end
       write_annotation(object) if pretty
       output
+    end
+
+    QUOTED_STRING_REGEXP = %r{[^\w\./]}
+    private_constant :QUOTED_STRING_REGEXP
+
+    def write_string_quoted_if_necessary(object)
+      string = object.to_s
+      string =~ QUOTED_STRING_REGEXP ? write_quoted_string(string) : write_string(string)
     end
 
     def write_string(object)
